backend/requirements.txt
tiktoken
fastapi
uvicorn
pydantic
groq
sentence-transformers
faiss-cpu
numpy
PyPDF2
python-multipart